created a helper function to create the authorization header

+++ src/redux/thunks/deleteTodo.ts
@@ -1,6 +1,7 @@
 import {createAsyncThunk} from '@reduxjs/toolkit';
 import Config from 'react-native-config';
 import {RootState} from '../types/rootState.ts';
+import createAuthorizationHeader from '../../utils/createAuthorizationHeader.ts';
 
 export default createAsyncThunk(
   // This argument is the action name
@@ -10,15 +11,15 @@ export default createAsyncThunk(
   async (todoID: number, thunkAPI) => {
     const url = `${Config.API_URL}/todos/${todoID}`;
 
-    const {user} = thunkAPI.getState() as RootState;
+    const token = (thunkAPI.getState() as RootState).user.token;
 
-    const headers = {
-      Authorization: `Bearer ${user.token}`,
-    };
+    if (!token) {
+      return thunkAPI.rejectWithValue('Kein Token vorhanden');
+    }
 
     const response = await fetch(url, {
       method: 'DELETE',
-      headers,
+      headers: createAuthorizationHeader(token),
     });
 
     if (response.status !== 200) {
